+++ src/test/java/com/atlassian/db/replica/api/ConnectionStateTest.java
@@ -1,9 +1,8 @@
-package com.atlassian.db.replica.internal.state;
+package com.atlassian.db.replica.api;
 
-import com.atlassian.db.replica.api.DualConnection;
 import com.atlassian.db.replica.api.mocks.CircularConsistency;
 import com.atlassian.db.replica.api.mocks.ConnectionProviderMock;
-import com.atlassian.db.replica.spi.state.StateListener;
+import com.atlassian.db.replica.internal.state.StateListener;
 import com.google.common.collect.ImmutableList;
 import org.junit.Before;
 import org.junit.Test;
@@ -16,11 +15,11 @@ import static com.atlassian.db.replica.api.Queries.SELECT_FOR_UPDATE;
 import static com.atlassian.db.replica.api.Queries.SIMPLE_QUERY;
 import static com.atlassian.db.replica.api.mocks.CircularConsistency.permanentConsistency;
 import static com.atlassian.db.replica.api.mocks.CircularConsistency.permanentInconsistency;
-import static com.atlassian.db.replica.api.state.State.CLOSED;
-import static com.atlassian.db.replica.api.state.State.MAIN;
-import static com.atlassian.db.replica.api.state.State.NOT_INITIALISED;
-import static com.atlassian.db.replica.api.state.State.COMMITED_MAIN;
-import static com.atlassian.db.replica.api.state.State.REPLICA;
+import static com.atlassian.db.replica.internal.state.State.CLOSED;
+import static com.atlassian.db.replica.internal.state.State.COMMITED_MAIN;
+import static com.atlassian.db.replica.internal.state.State.MAIN;
+import static com.atlassian.db.replica.internal.state.State.NOT_INITIALISED;
+import static com.atlassian.db.replica.internal.state.State.REPLICA;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;