+++ CHANGELOG.md
@@ -23,6 +23,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.31...master
 
+### Removed
+- `DualConnection.Builder.stateListener()`
+- `StateListener`
+- `State`
+
 ## [0.1.31] - 2021-03-10
 [0.1.31]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.30...release-0.1.31
 
@@ -30,9 +35,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 -`RouteDecision#mustRunOnMain()` and `RouteDecision#willRunOnMain()`
 
 ### Fixed
-- clarified read/write Reasons 
+- clarified read/write Reasons
 
-### Removed 
+### Removed
 - `Reason.isWrite()`
 
 ## [0.1.30] - 2021-03-09