+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.31...master
+[Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-1.0.0...master
+
+## [1.0.0] - 2021-03-15
+[1.0.0]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.30...release-1.0.0
 
 ### Removed
 - `DualConnection.Builder.stateListener()`